feat(getIsTileValidCombination): init - split out from getOneMixSuitScore
