+++ src/services/getOneMixSuitScore.ts
@@ -1,44 +1,7 @@
 import * as constants from "../constants";
 
+import getIsTileValidCombination from "./getIsTileValidCombination";
 import getIsSuited from "./getIsSuited";
-
-const getIsSuitHelper = (tiles: Array<constants.TileType>): boolean => {
-  // has to be at least 3 tiles
-  if (!tiles.length) return true;
-  if (tiles.length < 3) return false;
-
-  // sort the tiles in place
-  tiles.sort();
-
-  let curr = tiles[0];
-
-  // try forming triplets
-  if (curr === tiles[1] && curr === tiles[2]) {
-    const copy = [...tiles];
-    copy.splice(0, 3);
-    if (getIsSuitHelper(copy)) return true;
-  }
-
-  // try forming chow
-  const [currSuit, currIdx] = curr.split(".");
-  const secondIdx = tiles.findIndex(
-    (itm) => itm === `${currSuit}.${Number(currIdx) + 1}`
-  );
-  const thirdIdx = tiles.findIndex(
-    (itm) => itm === `${currSuit}.${Number(currIdx) + 2}`
-  );
-  const copy = [...tiles];
-  if (secondIdx >= 0 || thirdIdx >= 0) {
-    // remove this three items
-    // note: has to remove in decending order such that the index is correct
-    copy.splice(thirdIdx, 1);
-    copy.splice(secondIdx, 1);
-    copy.splice(0, 1);
-    return getIsSuitHelper(copy);
-  }
-
-  return false;
-};
 
 const getOneMixSuitScore = (
   tiles: Array<constants.TileType>
@@ -80,7 +43,7 @@ const getOneMixSuitScore = (
     }
 
     // check if it is successful
-    if (getIsSuitHelper(copy))
+    if (getIsTileValidCombination(copy))
       return isMixed
         ? constants.TileScore.MixedOneSuit
         : constants.TileScore.AllOneSuit;
